requirements.txt
apturl==0.5.2
asgiref==3.4.1
asn1crypto==0.24.0
backports.zoneinfo==0.2.1
Brlapi==0.6.6
certifi==2022.12.7
chardet==3.0.4
charset-normalizer==2.0.12
click==6.7
colorama==0.3.7
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cupshelpers==1.0
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
Django==3.2.18
djangorestframework==3.14.0
drf-yasg==1.21.5
Faker==13.11.1
httplib2==0.9.2
idna==3.4
importlib-resources==5.4.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==2.0.1
netifaces==0.10.4
oauth==1.0.1
olefile==0.45.1
packaging==21.3
pexpect==4.2.1
Pillow==5.1.0
protobuf==3.0.0
psycopg==3.0.14
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==3.0.9
pyRFC3339==1.0
python-apt==1.6.6
python-dateutil==2.8.2
python-debian==0.1.32
pytz==2022.7.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.27.1
requests-unixsocket==0.1.5
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
SecretStorage==2.3.1
simplejson==3.13.2
six==1.16.0
sqlparse==0.4.3
system-service==0.3
systemd-python==234
typing-extensions==4.1.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.14
usb-creator==0.3.3
wadllib==1.3.2
xkit==0.0.0
zipp==3.6.0
zope.interface==4.3.2
